#4 rename: stock 폴더 상위로 이동
