+++ src/main/java/sample/market/infrastructure/stock/StockRepository.java
@@ -1,4 +1,4 @@
-package sample.market.infrastructure.product.stock;
+package sample.market.infrastructure.stock;
 
 import jakarta.persistence.LockModeType;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import sample.market.domain.product.stock.Stock;
+import sample.market.domain.stock.Stock;
 
 @Repository
 public interface StockRepository extends JpaRepository<Stock, Long> {
