+++ src/test/java/sample/market/domain/stock/StockServiceTest.java
@@ -1,4 +1,4 @@
-package sample.market.infrastructure.product.stock;
+package sample.market.domain.stock;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,17 +6,16 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.repository.query.Param;
-import org.springframework.transaction.annotation.Transactional;
 import sample.market.domain.product.Product;
 import sample.market.domain.product.ProductStore;
-import sample.market.domain.product.stock.Stock;
+import sample.market.domain.stock.Stock;
+import sample.market.domain.stock.StockService;
 import sample.market.domain.user.User;
 import sample.market.domain.user.UserStore;
+import sample.market.infrastructure.stock.StockRepository;
 
 @SpringBootTest
-@Transactional
-class StockRepositoryTest {
+class StockServiceTest {
     @Autowired
     private UserStore userStore;
 
@@ -26,9 +25,12 @@ class StockRepositoryTest {
     @Autowired
     private StockRepository stockRepository;
 
-    @DisplayName("productId로 상품의 재고를 조회한다.")
+    @Autowired
+    private StockService stockService;
+
+    @DisplayName("비관적 락 - productId로 조회한 상품의 재고를 1 감소시킨다.")
     @Test
-    void findByProductId() {
+    void decreaseWithPessimistic() {
         // given
         User seller = User.builder()
                 .email("email@email.com")
@@ -53,24 +55,25 @@ class StockRepositoryTest {
         product1.reserved();
         productStore.store(product1);
 
+        Long quantity = 100L;
         Stock stock = Stock.builder()
                 .productId(product1.getId())
-                .quantity(100L)
+                .quantity(quantity)
                 .build();
-        stockRepository.save(stock);
+        Stock savedStock = stockRepository.save(stock);
 
         // when
-        Stock getStock = stockRepository.findByProductId(product1.getId());
+        stockService.decreaseWithPessimistic(product1.getId());
+        Stock getStock = stockRepository.findById(savedStock.getId()).orElseThrow();
+
 
         // then
-        assertThat(getStock.getProductId()).isEqualTo(product1.getId());
-        assertThat(getStock.getQuantity()).isEqualTo(stock.getQuantity());
-
+        assertThat(getStock.getQuantity()).isEqualTo(quantity - 1);
     }
 
-    @DisplayName("비관적 락 - productId로 상품의 재고를 조회한다.")
+    @DisplayName("낙관적 락 - productId로 조회한 상품의 재고를 1 감소시킨다.")
     @Test
-    void findByIdWithPessimisticLock() {
+    void decreaseWithOptimistic() throws InterruptedException {
         // given
         User seller = User.builder()
                 .email("email@email.com")
@@ -95,61 +98,18 @@ class StockRepositoryTest {
         product1.reserved();
         productStore.store(product1);
 
+        Long quantity = 100L;
         Stock stock = Stock.builder()
                 .productId(product1.getId())
-                .quantity(100L)
+                .quantity(quantity)
                 .build();
-        stockRepository.save(stock);
+        Stock savedStock = stockRepository.save(stock);
 
         // when
-        Stock getStock = stockRepository.findByIdWithPessimisticLock(product1.getId());
+        stockService.decreaseWithOptimistic(product1.getId());
+        Stock getStock = stockRepository.findById(savedStock.getId()).orElseThrow();
 
         // then
-        assertThat(getStock.getProductId()).isEqualTo(product1.getId());
-        assertThat(getStock.getQuantity()).isEqualTo(stock.getQuantity());
+        assertThat(getStock.getQuantity()).isEqualTo(quantity - 1);
     }
-
-    @DisplayName("낙관적 락 - productId로 상품의 재고를 조회한다.")
-    @Test
-    void findByIdWithOptimisticLock() {
-        // given
-        User seller = User.builder()
-                .email("email@email.com")
-                .password("password")
-                .role("user")
-                .username("username")
-                .build();
-        User buyer = User.builder()
-                .email("email@email.com")
-                .password("password")
-                .role("user")
-                .username("username")
-                .build();
-        userStore.store(seller);
-        userStore.store(buyer);
-
-        Product product1 = Product.builder()
-                .price(1000)
-                .name("마스크")
-                .sellerId(seller.getId())
-                .build();
-        product1.reserved();
-        productStore.store(product1);
-
-        Stock stock = Stock.builder()
-                .productId(product1.getId())
-                .quantity(100L)
-                .build();
-        stockRepository.save(stock);
-
-        // when
-        Stock getStock = stockRepository.findByIdWithOptimisticLock(product1.getId());
-
-        // then
-        assertThat(getStock.getProductId()).isEqualTo(product1.getId());
-        assertThat(getStock.getQuantity()).isEqualTo(stock.getQuantity());
-    }
-
-
-
 }
