+++ src/test/java/sample/market/infrastructure/stock/StockRepositoryTest.java
@@ -1,4 +1,4 @@
-package sample.market.domain.product.stock;
+package sample.market.infrastructure.stock;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,14 +6,17 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
 import sample.market.domain.product.Product;
 import sample.market.domain.product.ProductStore;
+import sample.market.domain.stock.Stock;
 import sample.market.domain.user.User;
 import sample.market.domain.user.UserStore;
-import sample.market.infrastructure.product.stock.StockRepository;
+import sample.market.infrastructure.stock.StockRepository;
 
 @SpringBootTest
-class StockServiceTest {
+@Transactional
+class StockRepositoryTest {
     @Autowired
     private UserStore userStore;
 
@@ -23,12 +26,9 @@ class StockServiceTest {
     @Autowired
     private StockRepository stockRepository;
 
-    @Autowired
-    private StockService stockService;
-
-    @DisplayName("비관적 락 - productId로 조회한 상품의 재고를 1 감소시킨다.")
+    @DisplayName("productId로 상품의 재고를 조회한다.")
     @Test
-    void decreaseWithPessimistic() {
+    void findByProductId() {
         // given
         User seller = User.builder()
                 .email("email@email.com")
@@ -53,25 +53,24 @@ class StockServiceTest {
         product1.reserved();
         productStore.store(product1);
 
-        Long quantity = 100L;
         Stock stock = Stock.builder()
                 .productId(product1.getId())
-                .quantity(quantity)
+                .quantity(100L)
                 .build();
-        Stock savedStock = stockRepository.save(stock);
+        stockRepository.save(stock);
 
         // when
-        stockService.decreaseWithPessimistic(product1.getId());
-        Stock getStock = stockRepository.findById(savedStock.getId()).orElseThrow();
-
+        Stock getStock = stockRepository.findByProductId(product1.getId());
 
         // then
-        assertThat(getStock.getQuantity()).isEqualTo(quantity - 1);
+        assertThat(getStock.getProductId()).isEqualTo(product1.getId());
+        assertThat(getStock.getQuantity()).isEqualTo(stock.getQuantity());
+
     }
 
-    @DisplayName("낙관적 락 - productId로 조회한 상품의 재고를 1 감소시킨다.")
+    @DisplayName("비관적 락 - productId로 상품의 재고를 조회한다.")
     @Test
-    void decreaseWithOptimistic() throws InterruptedException {
+    void findByIdWithPessimisticLock() {
         // given
         User seller = User.builder()
                 .email("email@email.com")
@@ -96,18 +95,61 @@ class StockServiceTest {
         product1.reserved();
         productStore.store(product1);
 
-        Long quantity = 100L;
         Stock stock = Stock.builder()
                 .productId(product1.getId())
-                .quantity(quantity)
+                .quantity(100L)
                 .build();
-        Stock savedStock = stockRepository.save(stock);
+        stockRepository.save(stock);
 
         // when
-        stockService.decreaseWithOptimistic(product1.getId());
-        Stock getStock = stockRepository.findById(savedStock.getId()).orElseThrow();
+        Stock getStock = stockRepository.findByIdWithPessimisticLock(product1.getId());
 
         // then
-        assertThat(getStock.getQuantity()).isEqualTo(quantity - 1);
+        assertThat(getStock.getProductId()).isEqualTo(product1.getId());
+        assertThat(getStock.getQuantity()).isEqualTo(stock.getQuantity());
     }
+
+    @DisplayName("낙관적 락 - productId로 상품의 재고를 조회한다.")
+    @Test
+    void findByIdWithOptimisticLock() {
+        // given
+        User seller = User.builder()
+                .email("email@email.com")
+                .password("password")
+                .role("user")
+                .username("username")
+                .build();
+        User buyer = User.builder()
+                .email("email@email.com")
+                .password("password")
+                .role("user")
+                .username("username")
+                .build();
+        userStore.store(seller);
+        userStore.store(buyer);
+
+        Product product1 = Product.builder()
+                .price(1000)
+                .name("마스크")
+                .sellerId(seller.getId())
+                .build();
+        product1.reserved();
+        productStore.store(product1);
+
+        Stock stock = Stock.builder()
+                .productId(product1.getId())
+                .quantity(100L)
+                .build();
+        stockRepository.save(stock);
+
+        // when
+        Stock getStock = stockRepository.findByIdWithOptimisticLock(product1.getId());
+
+        // then
+        assertThat(getStock.getProductId()).isEqualTo(product1.getId());
+        assertThat(getStock.getQuantity()).isEqualTo(stock.getQuantity());
+    }
+
+
+
 }
